+++ packages/device_info/ios/Classes/DeviceInfoPlugin.m
@@ -3,6 +3,7 @@
 // found in the LICENSE file.
 
 #import "DeviceInfoPlugin.h"
+#import <sys/utsname.h>
 
 @implementation DeviceInfoPlugin
 + (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar>*)registrar {
@@ -16,6 +17,9 @@
 - (void)handleMethodCall:(FlutterMethodCall*)call result:(FlutterResult)result {
   if ([@"getIosDeviceInfo" isEqualToString:call.method]) {
     UIDevice* device = [UIDevice currentDevice];
+    struct utsname un;
+    uname(&un);
+
     result(@{
       @"name" : [device name],
       @"systemName" : [device systemName],
@@ -23,10 +27,29 @@
       @"model" : [device model],
       @"localizedModel" : [device localizedModel],
       @"identifierForVendor" : [[device identifierForVendor] UUIDString],
+      @"isPhysicalDevice" : [self isDevicePhysical],
+      @"utsname" : @{
+        @"sysname" : @(un.sysname),
+        @"nodename" : @(un.nodename),
+        @"release" : @(un.release),
+        @"version" : @(un.version),
+        @"machine" : @(un.machine),
+      }
     });
   } else {
     result(FlutterMethodNotImplemented);
   }
+}
+
+// return value is false if code is run on a simulator
+- (NSString*)isDevicePhysical {
+#if TARGET_OS_SIMULATOR
+  NSString* isPhysicalDevice = @"false";
+#else
+  NSString* isPhysicalDevice = @"true";
+#endif
+
+  return isPhysicalDevice;
 }
 
 @end