+++ packages/firebase_messaging/ios/Classes/FirebaseMessagingPlugin.m
@@ -28,17 +28,14 @@
 
 - (instancetype)initWithChannel:(FlutterMethodChannel *)channel {
   self = [super init];
+
   if (self) {
     _channel = channel;
     _resumingFromBackground = NO;
     if (![FIRApp defaultApp]) {
       [FIRApp configure];
     }
-    [FIRMessaging messaging].remoteMessageDelegate = self;
-    [[NSNotificationCenter defaultCenter] addObserver:self
-                                             selector:@selector(tokenRefreshNotification:)
-                                                 name:kFIRInstanceIDTokenRefreshNotification
-                                               object:nil];
+    [FIRMessaging messaging].delegate = self;
   }
   return self;
 }
@@ -81,37 +78,12 @@
   }
 }
 
-- (void)tokenRefreshNotification:(NSNotification *)notification {
-  NSString *refreshedToken = [[FIRInstanceID instanceID] token];
-
-  // Connect to FCM since connection may have failed when attempted before having a token.
-  [self connectToFcm];
-
-  [_channel invokeMethod:@"onToken" arguments:refreshedToken];
-}
-
 #if defined(__IPHONE_10_0) && __IPHONE_OS_VERSION_MAX_ALLOWED >= __IPHONE_10_0
 // Receive data message on iOS 10 devices while app is in the foreground.
 - (void)applicationReceivedRemoteMessage:(FIRMessagingRemoteMessage *)remoteMessage {
   [self didReceiveRemoteNotification:remoteMessage.appData];
 }
 #endif
-
-- (void)connectToFcm {
-  // Won't connect since there is no token
-  if (![[FIRInstanceID instanceID] token]) {
-    return;
-  }
-
-  // Disconnect previous FCM connection if it exists.
-  [[FIRMessaging messaging] disconnect];
-
-  [[FIRMessaging messaging] connectWithCompletion:^(NSError *_Nullable error) {
-    if (error != nil) {
-      NSLog(@"Unable to connect to FCM. %@", error);
-    }
-  }];
-}
 
 - (void)didReceiveRemoteNotification:(NSDictionary *)userInfo {
   if (_resumingFromBackground) {
@@ -132,13 +104,11 @@
 }
 
 - (void)applicationDidEnterBackground:(UIApplication *)application {
-  [[FIRMessaging messaging] disconnect];
   _resumingFromBackground = YES;
 }
 
 - (void)applicationDidBecomeActive:(UIApplication *)application {
   _resumingFromBackground = NO;
-  [self connectToFcm];
   // Clears push notifications from the notification center, with the
   // side effect of resetting the badge count. We need to clear notifications
   // because otherwise the user could tap notifications in the notification
@@ -147,8 +117,10 @@
   // user dismissed the notification center without tapping anything.
   // TODO(goderbauer): Revisit this behavior once we provide an API for managing
   // the badge number, or if we add support for running Dart in the background.
-  // Setting badgeNumber to 0 is a no-op (= notifications will not be cleared) if it is already 0,
-  // therefore the next line is setting it to 1 first before clearing it again to remove all
+  // Setting badgeNumber to 0 is a no-op (= notifications will not be cleared)
+  // if it is already 0,
+  // therefore the next line is setting it to 1 first before clearing it again
+  // to remove all
   // notifications.
   application.applicationIconBadgeNumber = 1;
   application.applicationIconBadgeNumber = 0;
@@ -175,6 +147,11 @@
     @"alert" : [NSNumber numberWithBool:notificationSettings.types & UIUserNotificationTypeAlert],
   };
   [_channel invokeMethod:@"onIosSettingsRegistered" arguments:settingsDictionary];
+}
+
+- (void)messaging:(nonnull FIRMessaging *)messaging
+    didRefreshRegistrationToken:(nonnull NSString *)fcmToken {
+  [_channel invokeMethod:@"onToken" arguments:fcmToken];
 }
 
 @end