+++ packages/firebase_auth/CHANGELOG.md
@@ -1,6 +1,23 @@
+## 0.3.2
+
+* Added support for changing user data
+
+## 0.3.1
+
+* Change GMS dependency to 11.+
+
+## 0.3.0
+
+* **Breaking Change**: Method FirebaseUser getToken was renamed to getIdToken.
+
+## 0.2.5
+
+* Added support for linkWithCredential with Google credential
+
 ## 0.2.4
 
-Added support for changing user data
+* Added support for `signInWithCustomToken`
+* Added `Stream<FirebaseUser> onAuthStateChanged` event to listen when the user change
 
 ## 0.2.3+1