+++ packages/google_sign_in/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.2
+
+* Support setting foregroundColor in the avatar.
+
+## 1.0.1
+
+* Change GMS dependency to 11.+
+
 ## 1.0.0
 
 * Make GoogleUserCircleAvatar fade profile image over the top of placeholder